requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
websockets>=13.0
pydantic>=2.9.0
python-dotenv>=1.0.0
numpy>=2.0.0
pandas>=2.2.0
scikit-learn>=1.5.0
scipy>=1.14.0
joblib>=1.4.0
httpx>=0.27.0
pytest>=8.0.0
aiofiles>=24.0.0
